requirements.txt
boto3==1.28.57
botocore==1.31.57
Django==4.2.4
Django==3.2.21
djangorestframework==3.14.0
python-dotenv==1.0.0
